Authorized QUestionPolicy for edit and destroy

+++ app/Http/Controllers/QuestionsController.php
@@ -36,13 +36,17 @@ class QuestionsController extends Controller
 
     public function edit(Question $question)
     {
-        return view('questions.edit', compact([
-            'question'
-        ]));
+        if ($this->authorize('update', $question)) {
+            return view('questions.edit', compact([
+                'question'
+            ]));
+        }
+        abort(403);
     }
 
     public function update(UpdateQuestionRequest $request, Question $question)
     {
+
         $question->update([
             'title' => $request->title,
             'body' => $request->body
@@ -54,9 +58,15 @@ class QuestionsController extends Controller
 
     public function destroy(Question $question)
     {
-        $question->delete();
-        session()->flash('success', 'Question has been deleted successfully!');
-        return redirect(route('questions.index'));
+
+        if ($this->authorize('delete', $question)) {
+            $question->delete();
+            session()->flash('success', 'Question has been deleted successfully!');
+            return redirect(route('questions.index'));
+
+        }
+
+        abort(403);
 
     }
